Fixed compiling errors.

+++ consensus.hpp
@@ -2391,10 +2391,10 @@ appendVcfRecord(std::string & out_string, std::string & out_string_pass, const S
         // TODO: fill // mul_contam_rate < add_contam_rate
         double ntfrac = (double)(nDP0 + 1) / (double)(tDP0 + 1);
         
-        double add_contam_phred = calc_binom_10log10_likeratio(add_contam_rate, (double)nAD0, (double)tAD0); // 0.2 max 0.02 min
+        double add_contam_phred = calc_binom_10log10_likeratio((double)add_contam_rate, (double)nAD0, (double)tAD0); // 0.2 max 0.02 min
         double mul_contam_phred = (nDP0 < tDP0 
-            ? calc_binom_10log10_likeratio(mul_contam_rate, (double)nAD0         , (double)tAD0 * ntfrac)
-            : calc_binom_10log10_likeratio(mul_contam_rate, (double)nAD0 / ntfrac, (double)tAD0)
+            ? calc_binom_10log10_likeratio((double)mul_contam_rate, (double)nAD0         , (double)tAD0 * ntfrac)
+            : calc_binom_10log10_likeratio((double)mul_contam_rate, (double)nAD0 / ntfrac, (double)tAD0)
         );
         double contam_phred = MIN(add_contam_phred, mul_contam_phred); // select most likely contam model
